requirements.txt
PyMuPDF4LLM==0.0.17
tiktoken==0.5.1
openai==1.13.3
python-dotenv==1.0.1